Add fsupdate content handler
This adds a new update handler to adu agent. There are a few differences
when compared to the swupdate handler which fsupdate is based on.

Fsupdate hast to distinguish between ff and af files, these selectors
have been added to:
- Install
- IsInstalled

The additional value is stored as a member variable of the handler
object, and gets initialized when the object is created. To do this,
there is a new constructor as well as a new call.

The cancel Action was changed to noop, because there is no way to cancel
FS-Update