Run FS-Update as root
Some functions of FS-Update have to be run as root, other than adu-shell
FS-Update dosn´t do this by itself. To achieve this, we simply call
it as root.

Only the sub process is run as root, so that the permissions of
adu-agent aren't changed for the next deployment.

Remove Redirection of stdout and stderr to get log output of FS Update
in the same terminal.

Add extra log to see what child process was started, and which
parameters are used.