Add UpdateVersionFile
Add function and callback for updateVersionFile.
Add constructor with updateType

After a successful Apply the version file hast to be updated. To be more
specific, the installation criteria provided by the import manifest will
be written into this file. This value is stored inside the workflowData
struct.

The implementation is based on the IsInstalled function, that was
already provided by Azure.

This function can be outsourced to FS-Update. The implementation there
would be simpler, because there is no need for a callback.