Add GetUpdateRebootState
The UpdateVersionFile function wasn't needed, because the new version is
placed inside the update file, and will be there as soon as the
installation succeeds.

GetUpdateRebootState executes the FS-Updater with -urs to get the
current reboot state and determine which action has to be taken.

Change the agent startup workflow to use the new urs value.