extensions:fsupdate-handler: Improve handling to call fs-updater
To create and call a child process special rights are required.
adu-shell already has this necessary rights and can call other
commands. So all additional process calls should go through
adu shell.
Most of the fs-updater calls using adu shell already but
in case of firmware or application version fs-updater was started
directly. For this fs-updater was configured to had same addtional
rights as adu shell. But this not recommended because of security.
(s. special rights SUID)
Add the function HandleExecuteAction which handles special calls of
fs-updater to get state of update, commit or update version.
In this case fs-updater returns a value and/or output string.
To get correct value add special output of the string to parse
the value in fsupdate_handler.